+++ ts/features/bonus/bpd/store/actions/index.ts
@@ -1,8 +1,10 @@
 import { BpdDetailsActions } from "./details";
 import { BpdIbanActions } from "./iban";
 import { BpdOnboardingActions } from "./onboarding";
+import { BpdPaymentMethod } from "./paymentMethods";
 
 export type BpdActions =
   | BpdOnboardingActions
   | BpdDetailsActions
-  | BpdIbanActions;
+  | BpdIbanActions
+  | BpdPaymentMethod;