+++ ts/features/wallet/onboarding/bancomat/saga/networking/index.ts
@@ -6,7 +6,11 @@ import { SagaCallReturnType } from "../../../../../../types/utils";
 import { PaymentManagerClient } from "../../../../../../api/pagopa";
 import { SessionManager } from "../../../../../../utils/SessionManager";
 import { PaymentManagerToken } from "../../../../../../types/pagopa";
-import { loadAbi, searchUserPans } from "../../store/actions";
+import {
+  addBancomatToWallet,
+  loadAbi,
+  searchUserPans
+} from "../../store/actions";
 
 // load all bancomat abi
 export function* handleLoadAbi(
@@ -91,5 +95,49 @@ export function* handleLoadPans(
       );
     }
     return yield put(searchUserPans.failure({ kind: "generic", value: e }));
+  }
+}
+
+// add pan to wallet
+export function* handleAddPan(
+  addPans: ReturnType<typeof PaymentManagerClient>["addPans"],
+  sessionManager: SessionManager<PaymentManagerToken>,
+  action: ActionType<typeof addBancomatToWallet.request>
+) {
+  try {
+    const addPansWithRefresh = sessionManager.withRefresh(
+      // add a card as an array of one element
+      addPans({ data: [action.payload] })
+    );
+    const addPansWithRefreshResult: SagaCallReturnType<typeof addPansWithRefresh> = yield call(
+      addPansWithRefresh
+    );
+    if (addPansWithRefreshResult.isRight()) {
+      if (addPansWithRefreshResult.value.status === 200) {
+        const wallets = addPansWithRefreshResult.value.value.data ?? [];
+        // search for the added bancomat.
+        const maybeWallet = fromNullable(
+          wallets.find(w => w.info.hashPan === action.payload.hpan)
+        );
+        if (maybeWallet.isSome()) {
+          yield put(
+            // success
+            addBancomatToWallet.success(maybeWallet.value)
+          );
+        } else {
+          throw new Error(
+            `cannot find added bancomat in wallets list response`
+          );
+        }
+      } else {
+        throw new Error(
+          `response status ${addPansWithRefreshResult.value.status}`
+        );
+      }
+    } else {
+      throw new Error(readableReport(addPansWithRefreshResult.value));
+    }
+  } catch (e) {
+    yield put(addBancomatToWallet.failure(e));
   }
 }