+++ ts/features/wallet/onboarding/bancomat/store/actions/index.ts
@@ -4,7 +4,8 @@ import {
   createStandardAction
 } from "typesafe-actions";
 import { AbiListResponse } from "../../../../../../../definitions/pagopa/bancomat/AbiListResponse";
-import { PatchedCard } from "../../../../../bonus/bpd/api/patchedTypes";
+import { Card } from "../../../../../../../definitions/pagopa/bancomat/Card";
+import { PatchedWalletV2 } from "../../../../../../types/pagopa";
 import { LoadPansError } from "../../saga/networking";
 
 /**
@@ -23,7 +24,17 @@ export const searchUserPans = createAsyncAction(
   "WALLET_ONBOARDING_BANCOMAT_LOAD_PANS_REQUEST",
   "WALLET_ONBOARDING_BANCOMAT_LOAD_PANS_SUCCESS",
   "WALLET_ONBOARDING_BANCOMAT_LOAD_PANS_FAILURE"
-)<string | undefined, ReadonlyArray<PatchedCard>, LoadPansError>();
+)<string | undefined, ReadonlyArray<Card>, LoadPansError>();
+
+/**
+ * The user select the current bancomat to add to the wallet
+ */
+export const addBancomatToWallet = createAsyncAction(
+  "WALLET_ONBOARDING_BANCOMAT_ADD_REQUEST",
+  "WALLET_ONBOARDING_BANCOMAT_ADD_SUCCESS",
+  "WALLET_ONBOARDING_BANCOMAT_ADD_FAILURE"
+)<Card, PatchedWalletV2, Error>();
+
 /**
  * The user choose to start the workflow to add a new bancomat to the wallet
  */
@@ -56,6 +67,7 @@ export const walletAddBancomatBack = createStandardAction(
 export type AbiActions =
   | ActionType<typeof loadAbi>
   | ActionType<typeof searchUserPans>
+  | ActionType<typeof addBancomatToWallet>
   | ActionType<typeof walletAddBancomatStart>
   | ActionType<typeof walletAddBancomatCompleted>
   | ActionType<typeof walletAddBancomatCancel>